requirements.txt
d4rl @ git+https://github.com/rail-berkeley/d4rl@d842aa194b416e564e54b0730d9f934e3e32f854
git+https://github.com/rlworkgroup/metaworld.git@a0009ed9a208ff9864a5c1368c04c273bb20dd06#egg=metaworld
absl-py==1.1.0
astor==0.8.1
astunparse==1.6.3
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chex==0.1.3
click==8.1.3
cloudpickle==1.6.0
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
dill==0.3.5.1
distrax==0.1.2
dm-acme==0.4.0
dm-env==1.5
dm-haiku==0.0.6
dm-launchpad==0.5.0
dm-reverb==0.7.0
dm-sonnet==2.0.0
dm-tree==0.1.7
fasteners==0.17.3
flatbuffers==2.0
flax==0.5.0
fonttools==4.33.3
future==0.18.2
gast==0.5.3
glfw==2.5.3
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
gym==0.19.0
gym-notices==0.0.7
h5py==3.7.0
imageio==2.19.3
importlib-metadata==4.11.4
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax==0.3.13
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jaxlib==0.3.10+cuda11.cudnn82
jmp==0.0.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
labmaze==1.0.5
libclang==14.0.1
lxml==4.9.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
mkl-fft==1.3.1
mkl-service==2.4.0
mock==4.0.3
msgpack==1.0.4
mujoco==2.2.0
mujoco-py==2.0.2.5
lockfile==0.12.2
numpy==1.22.4
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
Pillow==7.2.0
portpicker==1.5.2
promise==2.3
protobuf==3.19.1
psutil==5.9.1
pyasn1-modules==0.2.8
pybullet==3.2.5
pygame==2.1.2
pyglet==1.5.0
PyOpenGL==3.1.6
pyparsing==2.4.7
pycparser==2.21
python-dateutil==2.8.2
requests-oauthlib==1.3.1
rlax==0.1.2
rsa==4.8
scipy==1.8.1
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-datasets==4.4.0
tensorflow-estimator==2.8.0
tensorflow-probability==0.15.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toolz==0.11.2
tqdm==4.64.0
typing_extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0